kites/config/configstore: add Set method
