+++ go/src/koding/kites/config/configstore/export_test.go
@@ -1,0 +1,6 @@
+package configstore
+
+// SetFlatKeyValue exports setFlatKeyValue for testing purposes.
+func SetFlatKeyValue(m map[string]interface{}, key, value string) error {
+	return setFlatKeyValue(m, key, value)
+}
